+++ .prettierignore
@@ -2,6 +2,6 @@ node_modules
 dist
 temp
 .env
-exports/measurements/inputs/*
-exports/measurements/export-measurements.ts
+exports/measurement/inputs/*
+exports/measurement/export-measurement.ts
 .eslintignore
