+++ exports/measurement/validate-and-read-measurement.ts
